requirements.txt
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cleo==0.8.1
click==8.0.3
clikit==0.6.2
crashtest==0.3.1
cryptography==36.0.1
dataclasses==0.6
distlib==0.3.4
filelock==3.4.1
Flask==2.0.2
Flask-HTTPAuth==4.5.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-metadata==1.7.0
importlib-resources==5.4.0
itsdangerous==2.0.1
jeepney==0.7.1
Jinja2==3.0.3
keyring==21.8.0
lockfile==0.12.2
MarkupSafe==2.0.1
msgpack==1.0.3
numpy==1.22.0
packaging==20.9
passlib==1.7.4
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.8.2
platformdirs==2.4.0
poetry==1.1.12
poetry-core==1.0.7
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.3.0
pylev==1.4.0
pyparsing==3.0.6
requests==2.27.1
requests-toolbelt==0.9.1
SecretStorage==3.3.1
shellingham==1.4.0
six==1.16.0
SQLAlchemy==1.4.29
tomlkit==0.8.0
urllib3==1.26.7
virtualenv==20.13.0
webencodings==0.5.1
Werkzeug==2.0.2
WTForms==3.0.1
zipp==3.6.0
